meetman-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
eval_type_backport==0.2.2
frozenlist==1.6.0
google-auth==2.40.1
google-genai==1.2.0
gotrue==2.12.0
groq==0.12.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
inquirerpy==0.3.4
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain-core==0.3.61
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42
mistralai==1.5.0
multidict==6.4.3
multiprocess==0.70.14
mypy_extensions==1.1.0
nestd==0.3.1
numpy==2.1.3
openai==1.54.1
orjson==3.9.15
packaging==24.2
pfzy==0.3.4
pluggy==1.6.0
portalocker==2.10.1
postgrest==0.18.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.24.14
PyPDF2==3.0.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
qdrant-client==1.14.2
realtime==2.5.0
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
robyn==0.64.2
rsa==4.9.1
rustimport==1.3.4
scikit-learn==1.5.2
scipy==1.15.3
sentry-sdk==2.19.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvloop==0.19.0
watchdog==4.0.1
wcwidth==0.2.13
websockets==13.1
yarl==1.20.0
zstandard==0.23.0
